Finshed event driven twitterbot
